Move vision controllers to routes dir
